requirements.txt
langchain==0.1.1
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.24.0
